[clang_parser] fix clang_parser for LLVM 8+
LLVM changed the behavior of clang_Cursor_isAnonymous in [1]. The old
behavior would returned false for direct-acccess anonymous structs
within other structs, whereas LLVM 8 returns true. To circumvent this
behavior change among LLVM versions, we keep track of all structs
defined within a struct. We don't parse the substruct recursively (if we
do it might be parsed multiple times, and since we don't know yet if
this is a direct or indirect strucutre, we might parse it incorrectly),
instead we keep the cursor saved in a map. If this substruct is later
declared as an attribute of the supersctruct, that means we have a
direct-accessible struct. We remove it from our map, and we parse
recursively (parsing recursively the cursor pointing to the declaration
will effectively parse the struct definition).

After the first parse, any struct left in our map is an indirect struct.
Since we skipped the parsing stage for those, we need to run
`clang_visitChildren` again for each struct cursor we have saved. We
repeat this until there's no unvisited structs in the map. Keep in mind
that while visiting a new struct we might find more indirect structs.

Also add Travis jobs to test against LLVM and clang 8 on Ubuntu.

[1]: llvm/llvm-project@c05e6f4